// ClientApp/src/app/models/clientContact.model.ts
export class ClientContact {
  clientContactInfoId?: number;
  firstName: string;
  lastName: string;
  email: string;
  cell: string;
  tell: string;
  username: string;
  password: string;
  acquisitionNumber?: string;
  accessCode: string;
  imagePath: string;
  date?: Date;
  clientId: number;
}